requirements.txt
Flask==2.0.3
requests==2.27.1
python-dotenv==0.19.2
Flask-SocketIO==5.1.1
gunicorn==20.1.0
openai>=1.0.0
beautifulsoup4==4.10.0
pytz==2021.3
eventlet==0.33.0
Werkzeug==2.0.3
markupsafe==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.3
openrouter==0.2.0 